+++ familyhelper-clannad-sdk/src/main/java/com/dwarfeng/familyhelper/clannad/sdk/util/ServiceExceptionCodes.java
@@ -1,0 +1,45 @@
+package com.dwarfeng.familyhelper.clannad.sdk.util;
+
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+
+/**
+ * 服务异常代码。
+ *
+ * @author DwArFeng
+ * @since 1.1.1
+ */
+public final class ServiceExceptionCodes {
+
+    private static int EXCEPTION_CODE_OFFSET = 10000;
+
+    public static final ServiceException.Code PROFILE_NOT_EXISTS =
+            new ServiceException.Code(offset(0), "profile not exists");
+    public static final ServiceException.Code USER_NOT_EXISTS =
+            new ServiceException.Code(offset(30), "user not exists");
+
+    private static int offset(int i) {
+        return EXCEPTION_CODE_OFFSET + i;
+    }
+
+    /**
+     * 获取异常代号的偏移量。
+     *
+     * @return 异常代号的偏移量。
+     */
+    public static int getExceptionCodeOffset() {
+        return EXCEPTION_CODE_OFFSET;
+    }
+
+    /**
+     * 设置异常代号的偏移量。
+     *
+     * @param exceptionCodeOffset 指定的异常代号的偏移量。
+     */
+    public static void setExceptionCodeOffset(int exceptionCodeOffset) {
+        EXCEPTION_CODE_OFFSET = exceptionCodeOffset;
+    }
+
+    private ServiceExceptionCodes() {
+        throw new IllegalStateException("禁止实例化");
+    }
+}